pyproject.toml
[project]
name = "doc-ingestion-router"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Daniel Andreasen", email = "akda@aarhus.dk" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "html2text>=2025.4.15",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-mock>=3.14.1",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]

[project.scripts]
doc-ingestion-router = "doc_ingestion_router:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
